Harmonize notification configuration
This change makes the configuration of notifications consistent across
Liberty, Mitaka and Newton releases.

It also aligns the configuration between the controller and volume roles
by allowing to define the notitication topic(s) and messaging driver.

Finally it preserves backward-compatibility when the 'notification' parameter
is a boolean value instead of a dict.

Change-Id: Ib135f4274c251c314377ff6be915480cd2e543cc